main demo page
This has the draggable widget named "bob" with sub-widgets that appear next to widget. The sub-widget locations change depending on which quadrant of the screen you are in eg. top-left, top-right, etc...

I have not yet figured out how to attach images to a repository although they are not entirely important.